requirements.txt
streamlit==1.39.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
mediapipe==0.10.14
tensorflow==2.17.0
keras==3.4.1
Pillow==10.4.0
huggingface-hub==0.25.2
